Improve resolved from logic.
Improve resolved from logic.
