+++ components/org.wso2.carbon.identity.branding.preference.management.core/src/main/java/org/wso2/carbon/identity/branding/preference/management/core/model/BrandingPreference.java
@@ -70,31 +70,12 @@ public class BrandingPreference {
     }
 
     public ResolvedFrom getResolvedFrom() {
+
         return resolvedFrom;
     }
 
     public void setResolvedFrom(String type, String name) {
+
         this.resolvedFrom = new ResolvedFrom(type, name);
-    }
-
-    /**
-     * Inner class for the "resolvedFrom" object.
-     */
-    public static class ResolvedFrom {
-        private String type;
-        private String name;
-
-        public ResolvedFrom(String type, String name) {
-            this.type = type;
-            this.name = name;
-        }
-
-        public String getType() {
-            return type;
-        }
-
-        public String getName() {
-            return name;
-        }
     }
 }
