+++ components/org.wso2.carbon.identity.branding.preference.management.core/src/main/java/org/wso2/carbon/identity/branding/preference/management/core/model/CustomText.java
@@ -28,6 +28,7 @@ public class CustomText {
     private String screen;
     private String locale;
     private Object preference;
+    private ResolvedFrom resolvedFrom;
 
     public String getType() {
 
@@ -77,5 +78,15 @@ public class CustomText {
     public void setPreference(Object preference) {
 
         this.preference = preference;
+    }
+
+    public ResolvedFrom getResolvedFrom() {
+
+        return resolvedFrom;
+    }
+
+    public void setResolvedFrom(String type, String name) {
+
+        this.resolvedFrom = new ResolvedFrom(type, name);
     }
 }
