+++ components/org.wso2.carbon.identity.branding.preference.management.core/src/main/java/org/wso2/carbon/identity/branding/preference/management/core/model/ResolvedFrom.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.carbon.identity.branding.preference.management.core.model;
+
+/**
+ * A model class representing a resolved from.
+ */
+public class ResolvedFrom {
+
+    private String type;
+    private String name;
+
+    public ResolvedFrom(String type, String name) {
+
+        this.type = type;
+        this.name = name;
+    }
+
+    public String getType() {
+
+        return type;
+    }
+
+    public String getName() {
+
+        return name;
+    }
+}
